Made executable
The Misamino dll has been converted to a cli application, which
responds to commands in the specified format.

Previously dynamic library functions are now called directly from the application.

Some windows specific parts of the code and dll macros have been removed or adjusted
to make the code compilable with a standard GNU compiler.